Refactor tandem plugin to adhere to established formats
