--- src/perovskite_tandem_database/schema_sections/__init__.py
@@ -1,29 +1,0 @@
-from perovskite_solar_cell_database.schema_sections.add import Add
-from perovskite_solar_cell_database.schema_sections.backcontact import Backcontact
-from perovskite_solar_cell_database.schema_sections.cell import Cell
-from perovskite_solar_cell_database.schema_sections.encapsulation import Encapsulation
-from perovskite_solar_cell_database.schema_sections.eqe import EQE
-from perovskite_solar_cell_database.schema_sections.etl import ETL
-from perovskite_solar_cell_database.schema_sections.formula_normalizer import (
-    PerovskiteFormulaNormalizer,
-)
-from perovskite_solar_cell_database.schema_sections.htl import HTL
-from perovskite_solar_cell_database.schema_sections.ions.ion import Ion
-from perovskite_solar_cell_database.schema_sections.jv import JV, JVcurve
-from perovskite_solar_cell_database.schema_sections.module import Module
-from perovskite_solar_cell_database.schema_sections.outdoor import Outdoor
-from perovskite_solar_cell_database.schema_sections.perovskite import Perovskite
-from perovskite_solar_cell_database.schema_sections.perovskite_deposition import (
-    PerovskiteDeposition,
-)
-from perovskite_solar_cell_database.schema_sections.stabilised import Stabilised
-from perovskite_solar_cell_database.schema_sections.stability import Stability
-from perovskite_solar_cell_database.schema_sections.substrate import Substrate
-from perovskite_solar_cell_database.schema_sections.utils import (
-    add_band_gap,
-    add_solar_cell,
-)
-
-# from perovskite_solar_cell_database.schema_sections.ref import Ref
-from .ref import Reference
-from .tandem import General
